Now I have some connectivity, but there are 2 issues
one larger than the other.
The simple issue is that the status bar is not registering that the
device has been connected.

The larger issue is that something seems to be preventing the discovery
from taking place after the connection is established. Not sure if this
is a flaw in functional ability of android (me thinking I could get rssi
of something already connected) or if I am doing something wrong. Will
likely need to create a branch to do some tests on explicitly
rssi/discovery