Moving old android project.
I don't want this branch to delete too much just yet...
but the organization is not favorable
for multiple android apps. The old app should serve
as a template for RSSI and bluetooth items. But this
branch is about building a test app to practice with
the GPS. So... thinking of the BT one as already merged
into where I currently am... this file organization makes
a bit more sense.